Remove useless events (#31161)
* Remove useless events

* Remove useless events
