+++ features/encrypt/core/src/main/java/org/apache/shardingsphere/encrypt/rule/EncryptRule.java
@@ -25,7 +25,6 @@ import org.apache.shardingsphere.encrypt.exception.metadata.EncryptTableNotFound
 import org.apache.shardingsphere.encrypt.exception.metadata.MismatchedEncryptAlgorithmTypeException;
 import org.apache.shardingsphere.encrypt.rule.attribute.EncryptTableMapperRuleAttribute;
 import org.apache.shardingsphere.encrypt.spi.EncryptAlgorithm;
-import org.apache.shardingsphere.infra.algorithm.core.config.AlgorithmConfiguration;
 import org.apache.shardingsphere.infra.exception.core.ShardingSpherePreconditions;
 import org.apache.shardingsphere.infra.rule.attribute.RuleAttributes;
 import org.apache.shardingsphere.infra.rule.scope.DatabaseRule;
@@ -36,6 +35,7 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * Encrypt rule.
@@ -56,20 +56,13 @@ public final class EncryptRule implements DatabaseRule {
         this.databaseName = databaseName;
         configuration = ruleConfig;
         tables = new LinkedHashMap<>();
-        Map<String, EncryptAlgorithm> encryptors = createEncryptors(ruleConfig);
+        Map<String, EncryptAlgorithm> encryptors = ruleConfig.getEncryptors().entrySet().stream()
+                .collect(Collectors.toMap(Entry::getKey, entry -> TypedSPILoader.getService(EncryptAlgorithm.class, entry.getValue().getType(), entry.getValue().getProps())));
         for (EncryptTableRuleConfiguration each : ruleConfig.getTables()) {
             each.getColumns().forEach(columnRuleConfig -> checkEncryptorType(columnRuleConfig, encryptors));
             tables.put(each.getName().toLowerCase(), new EncryptTable(each, encryptors));
         }
         attributes = new RuleAttributes(new EncryptTableMapperRuleAttribute(ruleConfig.getTables()));
-    }
-    
-    private Map<String, EncryptAlgorithm> createEncryptors(final EncryptRuleConfiguration ruleConfig) {
-        Map<String, EncryptAlgorithm> result = new LinkedHashMap<>(ruleConfig.getEncryptors().size(), 1F);
-        for (Entry<String, AlgorithmConfiguration> entry : ruleConfig.getEncryptors().entrySet()) {
-            result.put(entry.getKey(), TypedSPILoader.getService(EncryptAlgorithm.class, entry.getValue().getType(), entry.getValue().getProps()));
-        }
-        return result;
     }
     
     private void checkEncryptorType(final EncryptColumnRuleConfiguration columnRuleConfig, final Map<String, EncryptAlgorithm> encryptors) {
