+++ mode/core/src/main/java/org/apache/shardingsphere/mode/manager/context/ConfigurationContextManager.java
@@ -155,6 +155,11 @@ public final class ConfigurationContextManager {
      * @param ruleConfig rule configurations
      */
     public synchronized void alterRuleConfiguration(final String databaseName, final RuleConfiguration ruleConfig) {
+        // TODO add feature for partly refresh
+        // 1. Judge if impl partly interface
+        // 2. compare diff with current and ruleConfig
+        // 3. Do partly refresh
+        // 4. return
         try {
             ShardingSphereDatabase database = metaDataContexts.get().getMetaData().getDatabase(databaseName);
             Collection<ShardingSphereRule> rules = new LinkedList<>(database.getRuleMetaData().getRules());
@@ -176,6 +181,11 @@ public final class ConfigurationContextManager {
      * @param ruleConfig rule configurations
      */
     public synchronized void dropRuleConfiguration(final String databaseName, final RuleConfiguration ruleConfig) {
+        // TODO add feature for partly refresh
+        // 1. Judge if impl partly interface
+        // 2. compare diff with current and ruleConfig
+        // 3. Remove to be removed partial config
+        // 4. return
         try {
             ShardingSphereDatabase database = metaDataContexts.get().getMetaData().getDatabase(databaseName);
             Collection<ShardingSphereRule> rules = new LinkedList<>(database.getRuleMetaData().getRules());
